+++ src/main/java/onlyoffice/sdk/service/config/ConfigServiceImpl.java
@@ -19,6 +19,7 @@
 package onlyoffice.sdk.service.config;
 
 import com.atlassian.jira.component.ComponentAccessor;
+import com.atlassian.jira.issue.attachment.Attachment;
 import com.atlassian.jira.security.JiraAuthenticationContext;
 import com.atlassian.jira.user.ApplicationUser;
 import com.onlyoffice.manager.document.DocumentManager;
@@ -29,6 +30,7 @@ import com.onlyoffice.model.common.User;
 import com.onlyoffice.model.documenteditor.Config;
 import com.onlyoffice.model.documenteditor.config.document.Permissions;
 import com.onlyoffice.model.documenteditor.config.document.Type;
+import com.onlyoffice.model.documenteditor.config.editorconfig.Customization;
 import com.onlyoffice.model.documenteditor.config.editorconfig.Mode;
 import com.onlyoffice.model.documenteditor.config.editorconfig.customization.Anonymous;
 import com.onlyoffice.service.documenteditor.config.DefaultConfigService;
@@ -66,15 +68,22 @@ public class ConfigServiceImpl extends DefaultConfigService {
     public Permissions getPermissions(final String fileId) {
         JiraAuthenticationContext jiraAuthenticationContext = ComponentAccessor.getJiraAuthenticationContext();
         ApplicationUser user = jiraAuthenticationContext.getLoggedInUser();
+        Long attachmentId = Long.parseLong(fileId);
+        Attachment attachment = attachmentUtil.getAttachment(attachmentId);
 
         String fileName = getDocumentManager().getDocumentName(fileId);
 
-        Boolean editPermission = attachmentUtil.checkAccess(Long.parseLong(fileId), user, true);
+        Boolean editPermission = attachmentUtil.checkAccess(attachmentId, user, true);
         Boolean isEditable = super.getDocumentManager().isEditable(fileName);
+
+        boolean protect = user != null;
+        if ("owner".equals(getSettingsManager().getSetting("protect")) && user != null) {
+            protect = user.getKey().equals(attachment.getAuthorKey());
+        }
 
         return Permissions.builder()
                 .edit(editPermission && isEditable)
-                .protect(user != null)
+                .protect(protect)
                 .chat(user != null
                         && getSettingsManager().getSettingBoolean("customization.chat", true))
                 .build();
@@ -93,5 +102,20 @@ public class ConfigServiceImpl extends DefaultConfigService {
         } else {
             return super.getUser();
         }
+    }
+
+    @Override
+    public Customization getCustomization(final String fileId) {
+        Customization customization = super.getCustomization(fileId);
+
+        customization.setMacros(
+                getSettingsManager().getSettingBoolean("customization.macros", true)
+        );
+
+        customization.setPlugins(
+                getSettingsManager().getSettingBoolean("customization.plugins", true)
+        );
+
+        return customization;
     }
 }